graal fixes
