+++ resteasy-core/src/main/java/org/jboss/resteasy/microprofile/config/ServletContextConfigSource.java
@@ -1,22 +1,14 @@
 package org.jboss.resteasy.microprofile.config;
 
-import java.util.Collections;
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.Map;
-
-import javax.servlet.ServletContext;
-
 import org.eclipse.microprofile.config.spi.ConfigSource;
-import org.jboss.resteasy.core.ResteasyContext;
 
-public class ServletContextConfigSource implements ConfigSource {
-
+public class ServletContextConfigSource extends BaseServletConfigSource implements ConfigSource {
    private static final boolean SERVLET_AVAILABLE;
+   private static Class<?> clazz = null;
    static {
-      Class<?> clazz = null;
       try {
-         clazz = Class.forName(ServletContext.class.getName());
+         clazz = Class.forName("javax.servlet.ServletContext");
+         clazz = Class.forName("org.jboss.resteasy.microprofile.config.ServletContextConfigSourceImpl");
       }
       catch (Throwable e)
       {
@@ -25,57 +17,8 @@ public class ServletContextConfigSource implements ConfigSource {
       SERVLET_AVAILABLE = clazz != null;
    }
 
-   private volatile String name;
-
-   @Override
-   public Map<String, String> getProperties() {
-      if (!SERVLET_AVAILABLE) {
-         return Collections.<String, String>emptyMap();
-      }
-      ServletContext context = ResteasyContext.getContextData(ServletContext.class);
-      if (context == null) {
-         return Collections.<String, String>emptyMap();
-      }
-      Map<String, String> map = new HashMap<String, String>();
-      Enumeration<String> keys = context.getInitParameterNames();
-      if (keys != null) {
-         while (keys.hasMoreElements())
-         {
-            String key = keys.nextElement();
-            map.put(key, context.getInitParameter(key));
-         }
-      }
-      return map;
-   }
-
-   @Override
-   public String getValue(String propertyName) {
-      if (!SERVLET_AVAILABLE) {
-         return null;
-      }
-      ServletContext context = ResteasyContext.getContextData(ServletContext.class);
-      if (context == null) {
-         return null;
-      }
-      return context.getInitParameter(propertyName);
-   }
-
-   @Override
-   public String getName() {
-      if (name == null) {
-         synchronized(this) {
-            if (name == null) {
-               if (!SERVLET_AVAILABLE) {
-                  name = toString();
-               }
-               ServletContext servletContext = ResteasyContext.getContextData(ServletContext.class);
-               StringBuilder sb = new StringBuilder();
-               name = sb.append(servletContext != null ? servletContext.getServletContextName() : null)
-                     .append(":ServletContextConfigSource").toString();
-            }
-         }
-      }
-      return name;
+   public ServletContextConfigSource() {
+      super(SERVLET_AVAILABLE, clazz);
    }
 
    @Override
