+++ resteasy-core/src/main/java/org/jboss/resteasy/plugins/interceptors/GZIPDecodingInterceptor.java
@@ -17,7 +17,7 @@ import javax.ws.rs.ext.ReaderInterceptorContext;
 
 import org.jboss.resteasy.core.ResteasyContext;
 import org.jboss.resteasy.core.interception.jaxrs.ServerReaderInterceptorContext;
-import org.jboss.resteasy.plugins.server.ResteasyContextParameters;
+import org.jboss.resteasy.plugins.server.servlet.ResteasyContextParameters;
 import org.jboss.resteasy.resteasy_jaxrs.i18n.LogMessages;
 import org.jboss.resteasy.resteasy_jaxrs.i18n.Messages;
 import org.jboss.resteasy.spi.ResteasyConfiguration;
