+++ testsuite/integration-tests/src/test/java/org/jboss/resteasy/test/providers/jackson2/PreferJacksonOverJsonBClientTest.java
@@ -4,7 +4,7 @@ import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.logging.Logger;
-import org.jboss.resteasy.plugins.server.ResteasyContextParameters;
+import org.jboss.resteasy.plugins.server.servlet.ResteasyContextParameters;
 import org.jboss.resteasy.test.providers.jackson2.resource.MyEntity;
 import org.jboss.resteasy.test.providers.jackson2.resource.PreferJacksonOverJsonBClientResource;
 import org.jboss.resteasy.utils.PortProviderUtil;
