+++ testsuite/integration-tests/src/test/java/org/jboss/resteasy/test/tracing/OnDemandTracingTest.java
@@ -3,7 +3,7 @@ package org.jboss.resteasy.test.tracing;
 import org.jboss.arquillian.container.test.api.OperateOnDeployment;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.jboss.resteasy.plugins.server.ResteasyContextParameters;
+import org.jboss.resteasy.plugins.server.servlet.ResteasyContextParameters;
 import org.jboss.resteasy.tracing.api.RESTEasyTracing;
 import org.jboss.resteasy.spi.HttpResponseCodes;
 import org.junit.Assert;
