+++ testsuite/integration-tests/src/test/java/org/jboss/resteasy/test/validation/cdi/ApplicationScopeValidationTest.java
@@ -15,7 +15,7 @@ import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.resteasy.api.validation.ViolationReport;
 import org.jboss.resteasy.category.NotForForwardCompatibility;
-import org.jboss.resteasy.plugins.server.ResteasyContextParameters;
+import org.jboss.resteasy.plugins.server.servlet.ResteasyContextParameters;
 import org.jboss.resteasy.test.validation.cdi.resource.ApplicationScopeIRestServiceAppScoped;
 import org.jboss.resteasy.test.validation.cdi.resource.ApplicationScopeIRestServiceReqScoped;
 import org.jboss.resteasy.test.validation.cdi.resource.ApplicationScopeMyDto;
